requirements.txt
discord.py
yfinance
investpy
pandas
requests
google-generativeai
Flask
